[update]検索窓微修正(検索できない)
