--- app/assets/stylesheets/search.scss
@@ -1,3 +1,0 @@
-// Place all the styles related to the search controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: http://sass-lang.com/
