--- app/helpers/search_helper.rb
@@ -1,2 +1,0 @@
-module SearchHelper
-end
